requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
boto3==1.42.21
botocore==1.42.21
bs4==0.0.2
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
geopandas==1.1.1
git-filter-repo==2.47.0
google-auth==2.40.3
google-genai==1.24.0
googlemaps==4.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
kiwisolver==1.4.9
msgpack==1.1.2
narwhals==2.5.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pgeocode==0.5.0
pillow==11.3.0
plotly==6.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.11.7
pydantic_core==2.33.2
pyogrio==0.11.0
pyparsing==3.2.4
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.4
rsa==4.9.1
s3transfer==0.16.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uk_postcodes_parsing==1.2.1
urllib3==2.5.0
websockets==15.0.1
